⬆️ Last fixes, version and changelog update
* Set user identity provider in updateUser console method

* Set minimal web version allowed to new version (breaking changes for console user with console urls)

* Clean database on node

* Fix var names

* Update console names and api

* Change font to TTNorms **PRO** and fix emoji position

* Set language preferences

* Clear input on click button

* Fix company name in side bar

* Fix hidden message button

* FIx notification subscriber

* Fix Observables

* Add version changelogs

+++ twake/frontend/src/app/components/ScrollHiddenComponents/ScrollWithHiddenComponents.tsx
@@ -9,6 +9,7 @@ type PropsType = {
   scrollTopComponent?: JSX.Element;
   scrollBottomComponent?: JSX.Element;
   tag: string;
+  disabled?: boolean;
 };
 
 const ScrollWithHiddenComponents: FC<PropsType> = ({
@@ -16,14 +17,18 @@ const ScrollWithHiddenComponents: FC<PropsType> = ({
   scrollTopComponent,
   scrollBottomComponent,
   tag,
+  disabled,
 }) => {
   const ref = useRef<any>(null);
   const service = HiddenNotificationService.get(tag);
 
   useEffect(() => {
+    if (disabled) {
+      return;
+    }
     service.setScroller(ref.current.firstChild);
     return () => service.removeScroller();
-  }, [tag, ref]);
+  }, [tag, ref, disabled]);
 
   const beacon: number[] = service.useWatcher(() => [
     service.state.beaconTop.length,
